feat: #39 array 유틸함수 파일단위로 분리
