+++ src/utils/array/count.ts
@@ -1,0 +1,3 @@
+export const count = <T>(arr: readonly T[]) => {
+  return arr.length;
+};
