+++ src/utils/array/first.ts
@@ -1,0 +1,3 @@
+export const first = <T>(array: readonly T[]): T | undefined => {
+  return array[0];
+};
