+++ src/utils/array/last.ts
@@ -1,0 +1,3 @@
+export const last = <T>(array: readonly T[]): T | undefined => {
+  return array[array.length - 1];
+};
